+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- `shortCircuitCondition` detector: PR [#202](https://github.com/nowarp/misti/pull/202)
+
+### Added
+- `ShortCircuitCondition` detector: PR [#202](https://github.com/nowarp/misti/pull/202)
+
 ## [0.5.0] - 2024-10-31
 
 ### Added